requirements.txt
gpiozero
lgpio
luma.oled
Pillow
